refactor: tackle PR237 review mediums
Addressing NeOMakinG's review on #237:

- dedupe `normalizeMemberPool` / `isNonZero` into `memberPool.ts`
  + shared `types.ts` (position.ts and positions.ts were byte-identical
  except for `lastAddHeight` — extracted to prevent drift)
- `lockup.ts` now routes through `getThorchainMimir()` from validation.ts
  instead of its own `/thorchain/mimir` round-trip (dedupe fetches)
- `math.ts` `getLiquidityUnits` docstring now calls out the slip-adjust
  approximation so UX copy doesn't lie on asymmetric adds
- drop stale `./chains/cosmos/thor/lp/affiliate` subpath export from
  `package.json` (affiliate.ts is gone, consumers would get
  module-not-found on publish)
- re-export `ThorchainLpPosition` from `position.ts` so the dedicated
  `lp/position` subpath keeps resolving the type (codex-review caught
  this one on the refactor)
- lockup test now matches the shared mimir error shape
