+++ packages/core/chain/chains/cosmos/thor/lp/position.ts
@@ -3,82 +3,11 @@ import { cosmosRpcUrl } from '@vultisig/core-chain/chains/cosmos/cosmosRpcUrl'
 import { HttpResponseError } from '@vultisig/lib-utils/fetch/HttpResponseError'
 import { queryUrl } from '@vultisig/lib-utils/query/queryUrl'
 
+import { isNonZeroBaseUnit, normalizeMemberPool } from './memberPool'
 import { assertValidPoolId, thorchainMidgardBaseUrl } from './pools'
+import type { RawMemberResponse, ThorchainLpPosition } from './types'
 
-/**
- * Subset of a Midgard `/v2/member/{address}` pool entry. Midgard returns
- * many more fields per position — we keep just the ones the agent stack
- * surfaces in chat (units, added amounts, pending state, last-add timestamp).
- */
-export type ThorchainLpPosition = {
-  pool: string
-  liquidityUnits: string
-  runeAdded: string
-  assetAdded: string
-  runePending: string
-  assetPending: string
-  runeAddress: string
-  assetAddress: string
-  /**
-   * Unix seconds of the last add. Midgard returns this; thornode does
-   * not, so positions sourced from the thornode fallback leave this as
-   * `'0'` and populate `lastAddHeight` instead.
-   */
-  dateLastAdded: string
-  /**
-   * THORChain block height of the last add. Populated from thornode's
-   * `last_add_height` when the position comes from the thornode
-   * fallback path. Empty string when the position comes from Midgard
-   * (which does not expose block height on `/v2/member`). Either field
-   * is sufficient to gate the 1h lockup check — `dateLastAdded` via a
-   * wall-clock window, or `lastAddHeight` via a block-count window.
-   */
-  lastAddHeight: string
-  /**
-   * True when either side has a non-zero pending amount. Common for
-   * asymmetric adds that THORChain has not yet credited.
-   */
-  isPending: boolean
-}
-
-type RawMemberPool = {
-  pool?: string
-  liquidityUnits?: string
-  runeAdded?: string
-  assetAdded?: string
-  runePending?: string
-  assetPending?: string
-  runeAddress?: string
-  assetAddress?: string
-  dateLastAdded?: string
-}
-
-type RawMemberResponse = {
-  pools?: RawMemberPool[]
-}
-
-const isNonZero = (s: string | undefined): boolean => {
-  if (!s) return false
-  try {
-    return BigInt(s) > 0n
-  } catch {
-    return false
-  }
-}
-
-const normalizeMemberPool = (raw: RawMemberPool): ThorchainLpPosition => ({
-  pool: raw.pool ?? '',
-  liquidityUnits: raw.liquidityUnits ?? '0',
-  runeAdded: raw.runeAdded ?? '0',
-  assetAdded: raw.assetAdded ?? '0',
-  runePending: raw.runePending ?? '0',
-  assetPending: raw.assetPending ?? '0',
-  runeAddress: raw.runeAddress ?? '',
-  assetAddress: raw.assetAddress ?? '',
-  dateLastAdded: raw.dateLastAdded ?? '0',
-  lastAddHeight: '',
-  isPending: isNonZero(raw.runePending) || isNonZero(raw.assetPending),
-})
+export type { ThorchainLpPosition } from './types'
 
 export type GetThorchainLpPositionInput = {
   /** bech32 thor1... address (the RUNE side of the position). */
@@ -88,13 +17,6 @@ export type GetThorchainLpPositionInput = {
 }
 
 /**
- * Fetch a single LP position from Midgard.
- *
- * Returns `null` when:
- *   - the address has no positions at all (Midgard returns 404), or
- *   - the address has positions but none in the requested pool.
- */
-/**
  * Detect a 404 from queryUrl by reading the typed `HttpResponseError.status`
  * field that `@vultisig/lib-utils` `assertFetchResponse` now throws. The
  * previous implementation string-matched the error message; that worked
@@ -103,6 +25,13 @@ export type GetThorchainLpPositionInput = {
 const isMidgardNotFoundError = (err: unknown): boolean =>
   err instanceof HttpResponseError && err.status === 404
 
+/**
+ * Fetch a single LP position from Midgard.
+ *
+ * Returns `null` when:
+ *   - the address has no positions at all (Midgard returns 404), or
+ *   - the address has positions but none in the requested pool.
+ */
 export const getThorchainLpPosition = async ({
   thorAddress,
   pool,
@@ -174,9 +103,9 @@ export const getThorchainLpPositionFromThornode = async ({
   // Thornode always returns the endpoint with zeroed fields after a full
   // withdraw, so explicitly guard "everything is zero" as "no position".
   if (
-    !isNonZero(units) &&
-    !isNonZero(pendingRune) &&
-    !isNonZero(pendingAsset)
+    !isNonZeroBaseUnit(units) &&
+    !isNonZeroBaseUnit(pendingRune) &&
+    !isNonZeroBaseUnit(pendingAsset)
   ) {
     return null
   }
@@ -201,6 +130,7 @@ export const getThorchainLpPositionFromThornode = async ({
       typeof raw.last_add_height === 'number'
         ? String(raw.last_add_height)
         : '',
-    isPending: isNonZero(pendingRune) || isNonZero(pendingAsset),
+    isPending:
+      isNonZeroBaseUnit(pendingRune) || isNonZeroBaseUnit(pendingAsset),
   }
 }
